+++ packages/xf-fs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @xfcfam/xf-fs
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @xfcfam/xf@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes