+++ packages/xf-reactnative-fs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xfcfam/xf-reactnative-fs
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @xfcfam/xf@0.3.0
+  - @xfcfam/xf-fs@1.0.0
+
 ## 0.1.1
 
 ### Patch Changes