+++ packages/xf-rest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@xfcfam/xf-rest`
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @xfcfam/xf@0.3.0
+
 ## 0.3.0
 
 ### Minor Changes